requirements.txt
fastapi
uvicorn
streamlit
langchain
openai
faiss-cpu
requests
feedparser
